Remove unused content
